AB#245: Move files to relevant folders
