+++ anet-dictionary.yml
@@ -1406,6 +1406,3 @@ dashboards:
   - label: decisives
     data: /data/dashboards/decisives.json
     type: decisives
-  - label: process board
-    type: board
-    data: /data/dashboards/process.json
