+++ client/src/components/DiagramNode.js
