+++ client/src/pages/Routing.js
@@ -1,4 +1,5 @@
 import AppContext from "components/AppContext"
+import BoardDashboard from "components/BoardDashboard"
 import AuthorizationGroupEdit from "pages/admin/authorizationgroup/Edit"
 import AuthorizationGroupNew from "pages/admin/authorizationgroup/New"
 import AuthorizationGroupShow from "pages/admin/authorizationgroup/Show"
@@ -7,7 +8,6 @@ import AdminIndex from "pages/admin/Index"
 import MergeLocations from "pages/admin/MergeLocations"
 import MergePeople from "pages/admin/MergePeople"
 import MergePositions from "pages/admin/MergePositions"
-import BoardDashboard from "pages/dashboards/BoardDashboard"
 import DecisivesDashboard from "pages/dashboards/DecisivesDashboard"
 import KanbanDashboard from "pages/dashboards/KanbanDashboard"
 import GraphiQL from "pages/GraphiQL"
